requirements.txt
benzinga == 1.21
mongo == 0.2.0
pymongo == 4.10.1
pandas == 2.2.3
numpy == 2.1.2
scikit-learn == 1.5.2
python-dotenv == 1.0.1
joblib == 1.4.2
websocket-client == 1.8.0